build bare bones firmware image
Running "make" will produce a kernel image with an initrd linked in
named 'edgerouterlite.fw'. This file can be tftp booted after staging
to the next-server TFTP host as follows:

    Octeon ubnt_e100# dhcp
    Octeon ubnt_e100# tftp $loadaddr edgerouterlite.fw
    Octeon ubnt_e100# bootoctlinux $loadaddr coremask=0x3

The serial console is setup by default argument built into the kernel.
When booted a "press enter to activate this console" message will be
printed. Press enter to get a shell.